update Orchestration API
