+++ sharding-jdbc-orchestration/src/test/java/io/shardingsphere/jdbc/orchestration/yaml/masterslave/YamlOrchestrationMasterSlaveIntegrateTest.java
@@ -21,7 +21,6 @@ import com.google.common.base.Function;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import io.shardingsphere.jdbc.orchestration.api.OrchestrationMasterSlaveDataSourceFactory;
-import io.shardingsphere.jdbc.orchestration.api.util.OrchestrationDataSourceCloseableUtil;
 import io.shardingsphere.jdbc.orchestration.yaml.AbstractYamlDataSourceTest;
 import lombok.RequiredArgsConstructor;
 import org.junit.Test;
@@ -74,6 +73,6 @@ public class YamlOrchestrationMasterSlaveIntegrateTest extends AbstractYamlDataS
             stm.executeQuery("SELECT * FROM t_order_item");
             stm.executeQuery("SELECT * FROM t_config");
         }
-        OrchestrationDataSourceCloseableUtil.closeQuietly(dataSource);
+        OrchestrationMasterSlaveDataSourceFactory.closeQuietly(dataSource);
     }
 }
