+++ sharding-jdbc-orchestration/src/test/java/io/shardingsphere/jdbc/orchestration/yaml/sharding/YamlOrchestrationShardingWithMasterSlaveIntegrateTest.java
@@ -21,7 +21,6 @@ import com.google.common.base.Function;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import io.shardingsphere.jdbc.orchestration.api.OrchestrationShardingDataSourceFactory;
-import io.shardingsphere.jdbc.orchestration.api.util.OrchestrationDataSourceCloseableUtil;
 import io.shardingsphere.jdbc.orchestration.yaml.AbstractYamlDataSourceTest;
 import lombok.RequiredArgsConstructor;
 import org.junit.Test;
@@ -84,6 +83,6 @@ public class YamlOrchestrationShardingWithMasterSlaveIntegrateTest extends Abstr
             stm.executeQuery("SELECT * FROM t_order_item");
             stm.executeQuery("SELECT * FROM config");
         }
-        OrchestrationDataSourceCloseableUtil.closeQuietly(dataSource);
+        OrchestrationShardingDataSourceFactory.closeQuietly(dataSource);
     }
 }
